# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Enterprise HRBP Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /requisitions:
    get:
      operationId: listRequisitions
      tags: [requisitions]
      summary: List requisitions in scope
      parameters:
        - name: businessUnitId
          in: query
          required: false
          schema:
            type: integer
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Requisitions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Requisition"
    post:
      operationId: createRequisition
      tags: [requisitions]
      summary: Create a requisition
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequisitionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Requisition"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /requisitions/{id}:
    get:
      operationId: getRequisition
      tags: [requisitions]
      summary: Get a requisition
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Requisition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Requisition"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: updateRequisition
      tags: [requisitions]
      summary: Update a requisition
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequisitionUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Requisition"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteRequisition
      tags: [requisitions]
      summary: Delete a requisition
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /probation:
    get:
      operationId: listProbation
      tags: [probation]
      summary: List probation records in scope
      parameters:
        - name: businessUnitId
          in: query
          required: false
          schema:
            type: integer
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Probation records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Probation"
    post:
      operationId: createProbation
      tags: [probation]
      summary: Create a probation record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProbationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Probation"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /probation/{id}:
    get:
      operationId: getProbation
      tags: [probation]
      summary: Get a probation record
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Probation record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Probation"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: updateProbation
      tags: [probation]
      summary: Update a probation record
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProbationUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Probation"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteProbation
      tags: [probation]
      summary: Delete a probation record
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Requisition:
      type: object
      required:
        [id, title, businessUnitId, status, grade, openedDate, recruiter, createdAt, updatedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        businessUnitId:
          type: integer
        businessUnitName:
          type: ["string", "null"]
        status:
          type: string
          enum: [open, on_hold, filled, cancelled]
        grade:
          type: string
        openedDate:
          type: string
        targetCloseDate:
          type: ["string", "null"]
        filledDate:
          type: ["string", "null"]
        recruiter:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
    RequisitionInput:
      type: object
      required: [title, businessUnitId, status, grade, openedDate, recruiter]
      properties:
        title:
          type: string
          minLength: 1
        businessUnitId:
          type: integer
        status:
          type: string
          enum: [open, on_hold, filled, cancelled]
        grade:
          type: string
        openedDate:
          type: string
        targetCloseDate:
          type: ["string", "null"]
        filledDate:
          type: ["string", "null"]
        recruiter:
          type: string
    RequisitionUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        businessUnitId:
          type: integer
        status:
          type: string
          enum: [open, on_hold, filled, cancelled]
        grade:
          type: string
        openedDate:
          type: string
        targetCloseDate:
          type: ["string", "null"]
        filledDate:
          type: ["string", "null"]
        recruiter:
          type: string
...
    Probation:
      type: object
      required: [id, employeeId, businessUnitId, startDate, endDate, status, createdAt, updatedAt]
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        employeeName:
          type: ["string", "null"]
        businessUnitId:
          type: integer
        businessUnitName:
          type: ["string", "null"]
        startDate:
          type: string
        endDate:
          type: string
        status:
          type: string
          enum: [pending, passed, failed, extended]
        reviewDate:
          type: ["string", "null"]
        outcome:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    ProbationInput:
      type: object
      required: [employeeId, businessUnitId, startDate, endDate, status]
      properties:
        employeeId:
          type: integer
        businessUnitId:
          type: integer
        startDate:
          type: string
        endDate:
          type: string
        status:
          type: string
          enum: [pending, passed, failed, extended]
        reviewDate:
          type: ["string", "null"]
        outcome:
          type: ["string", "null"]
    ProbationUpdate:
      type: object
      properties:
        startDate:
          type: string
        endDate:
          type: string
        status:
          type: string
          enum: [pending, passed, failed, extended]
        reviewDate:
          type: ["string", "null"]
        outcome:
          type: ["string", "null"]
...